requirements.txt
# Optional: local tooling / IDE only. Runtime images use Dockerfiles.
apache-airflow==3.1.0
pandas>=2.1,<3
pyarrow>=14
statsforecast==1.7.7
joblib>=1.3
cloudpickle>=3
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2
requests>=2.31
streamlit>=1.39
